+++ orbit/pkg/table/extension_windows.go
@@ -7,7 +7,7 @@ import (
 
 	"github.com/fleetdm/fleet/v4/orbit/pkg/table/bitlocker_key_protectors"
 	cisaudit "github.com/fleetdm/fleet/v4/orbit/pkg/table/cis_audit"
-	mdmbridge "github.com/fleetdm/fleet/v4/orbit/pkg/table/mdm"
+	"github.com/fleetdm/fleet/v4/orbit/pkg/table/mdm_bridge"
 	"github.com/fleetdm/fleet/v4/orbit/pkg/table/windowsupdatetable"
 	"github.com/rs/zerolog/log"
 	"golang.org/x/sys/windows/registry"
@@ -32,7 +32,7 @@ func PlatformTables(_ PluginOpts) ([]osquery.OsqueryPlugin, error) {
 	}
 
 	if !windowsServer {
-		plugins = append(plugins, table.NewPlugin("mdm_bridge", mdmbridge.Columns(), mdmbridge.Generate))
+		plugins = append(plugins, table.NewPlugin("mdm_bridge", mdm_bridge.Columns(), mdm_bridge.Generate))
 	}
 
 	return plugins, nil
